requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pymongo==4.6.0
sentence-transformers==2.2.2
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
langchain-ollama==0.1.0
langchain-core==0.1.25
apscheduler==3.10.4
pydantic[email]==2.5.0
python-dotenv==1.0.0
slowapi==0.1.9
email-validator==2.1.0</content>
<parameter name="filePath">c:\Users\Enoch\Documents\GitHub\LLM\RAG (Daily_Questions)\requirements.txt